low battery detection test

+++ src/draw.cpp
@@ -22,7 +22,7 @@ static uint8_t add_sprite_entry(draw_sprite_entry* entry, uint8_t ci,
     {
         f += d * 2;
         if(state == STATE_MAP)
-            f += ((nframe >> 2) & 3);
+            f += (((uint8_t)nframe >> 2) & 3);
     }
     entry->addr = SPRITES_IMG;
     entry->frame = f;
@@ -63,7 +63,7 @@ void draw_sprites()
     // player sprite
     {
         uint8_t f = pdir * 4;
-        if(pmoving) f += ((nframe >> 2) & 3);
+        if(pmoving) f += (((uint8_t)nframe >> 2) & 3);
         entries[n++] = {PLAYER_IMG, f, 64 - 8, 32 - 8 - 4};
     }
 
@@ -89,7 +89,7 @@ void draw_sprites()
 void draw_player()
 {
     uint8_t f = pdir * 4;
-    if(pmoving) f += ((nframe >> 2) & 3);
+    if(pmoving) f += (((uint8_t)nframe >> 2) & 3);
     platform_fx_drawplusmask(64 - 8, 32 - 8 - 4, PLAYER_IMG, f, 16, 16);
 }
